automate vanilla registry extraction on linux
* Adding a simple check if someone has read the readme

* semi automate the registeries generation. still requires manually downloading server.jar for eula purposes

* The mods have spoken: no spelling mistakes allowed.

* mention extract_registries script in readme

---------

+++ build.sh
@@ -1,3 +1,11 @@
+#!/usr/bin/env bash
+
+if [ ! -f "include/registries.h" ]; then
+  echo "Error: 'include/registries.h' is missing."
+  echo "Please follow the 'Compilation' section of the README to generate it."
+  exit 1
+fi
+
 rm bareiron
 gcc src/*.c -O3 -Iinclude -o bareiron
 ./bareiron
